// demo/src/main/resources/frontend/angular-app/src/app/offer.ts
export interface Offer {
  id: number;
  name: string;
  continent: string;
  country: string;
  city: string;
  hotel: string;
  hotelStars: string;
  hotelDescription: string;
  airport: string;
  trip: string;
  startLocation: string;
  placeOfDestination: string;
  leaveDate: string;
  returnDate: string;
  howManyDays: string;
  typeOfFoodEnum: string;
  priceForAdult: string;
  priceForChild: string;
  promoted: string;
  placesForAdult: string;
  placesForChild: string;
}